Only send over the connection channel once
For some reason, upon reconnects, sending a second time would cause the ezsockets future to finish, and the disconnect procedure to start. Not sure why this is the case, but since we don't need the bridge ID and stream IDs a second time, we don't send them to the channel.

A more complete solution of only sending the required reconnect token on reconnect / proper callback layers in swb should be explored.